Renamed folder mov_cli to AniMov and all instances of it in the code
