+++ AniMov/websites/solar.py
